backend/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
PyJWT==1.7.1 
djangorestframework-jwt==1.11.0
Pillow==7.0.0